Kafka/KafkaClient/requirements.txt
confluent-kafka
numpy
pandas
pyspark
websocket-client
